% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_betas}
\alias{add_betas}
\title{Intermediary function to wrap text in "\\beta_{}"}
\usage{
add_betas(terms, nums)
}
\description{
Intermediary function to wrap text in "\\beta_{}"
}
\keyword{internal}
