% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_coefs.default}
\alias{add_coefs.default}
\title{Add coefficient values to the equation}
\usage{
\method{add_coefs}{default}(rhs, term, coef_digits)
}
\description{
Add coefficient values to the equation
}
\keyword{internal}
