% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_greek.default}
\alias{add_greek.default}
\title{Adds greek symbols to the equation}
\usage{
\method{add_greek}{default}(rhs, terms, greek = "beta", intercept = "alpha", raw_tex = FALSE)
}
\description{
Adds greek symbols to the equation
}
\keyword{internal}
