% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{add_se}
\alias{add_se}
\title{Add Standard Errors Below Coefficients}
\usage{
add_se(coef, model)
}
\arguments{
\item{coef}{character vector of model coefficients (from output of the function create_eq)}

\item{model}{a fitted model}
}
\value{
a character vector adding the errors beneath each term
}
\description{
Add Standard Errors Below Coefficients
}
