% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{add_tex_ital}
\alias{add_tex_ital}
\title{Wraps text in \code{\\text{}}}
\usage{
add_tex_ital(term, ital_vars)
}
\arguments{
\item{term}{A character to wrap in \code{\\text{}}}

\item{ital_vars}{Passed from \code{extract_eq}}
}
\value{
A character string
Add tex code to make string not italicized within an equation
}
\description{
Wraps text in \code{\\text{}}
}
\keyword{internal}
