% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{add_tex_ital_v}
\alias{add_tex_ital_v}
\title{Add tex code to make string not italicized within an equation for a vector
of strings}
\usage{
add_tex_ital_v(term_v, ital_vars)
}
\description{
Add tex code to make string not italicized within an equation for a vector
of strings
}
\keyword{internal}
