% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{add_tex_subscripts}
\alias{add_tex_subscripts}
\title{Wraps text in \code{_{}}}
\usage{
add_tex_subscripts(term)
}
\arguments{
\item{term}{A character string to TeXify}
}
\value{
A character string
Add tex code to make subscripts for a single string
}
\description{
Wraps text in \code{_{}}
}
\keyword{internal}
