% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{add_tex_subscripts_v}
\alias{add_tex_subscripts_v}
\title{Add tex code to make subscripts for a vector of strings}
\usage{
add_tex_subscripts_v(term_v)
}
\description{
Add tex code to make subscripts for a vector of strings
}
\keyword{internal}
