% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{align_eq}
\alias{align_eq}
\title{Applies align environment.}
\usage{
align_eq(.chrs, .align_env, ...)
}
\arguments{
\item{.chrs}{A list of characters. List of converted terms that will be modified and returned.}

\item{.align_env}{One of \code{c("aligned", "aligned*", "align", "align*")}.
TeX environment to wrap around equation.}

\item{...}{Other arguments that may be used by downstream functions.}
}
\value{
A \code{eq_term} classed list or list of \code{eq_term} classed lists.
}
\description{
Consolidation of align environment logic.
}
\seealso{
Other extensibility functions: 
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}

Other equatiomatic utility functions: 
\code{\link{wrap_eq}()}
}
\concept{equatiomatic utility functions}
\concept{extensibility functions}
\keyword{internal}
