% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{anno_greek}
\alias{anno_greek}
\title{Intermediary function to wrap text in \verb{\\\\beta_\{\}}}
\usage{
anno_greek(greek, nums, terms = NULL, raw_tex = FALSE)
}
\description{
Intermediary function to wrap text in \verb{\\\\beta_\{\}}
}
\keyword{internal}
