% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_property.R
\name{apply_property}
\alias{apply_property}
\alias{apply_estimate}
\title{Apply a function an \code{eq_term} object property.}
\usage{
apply_property(.x, .property, .f, .boolean_mask)

apply_estimate(.x, .f, .boolean_mask)
}
\arguments{
\item{.x}{An \code{eq_term} object or list of \code{eq_term} objects.}

\item{.property}{A character or vector of character. Name(s) of property(ies) to
be modified. If a vector, must be the same length as \code{.new}.}

\item{.f}{A function. The function applied to the current value in the
property.}

\item{.boolean_mask}{Optional. A vector of logicals (booleans) the same
length at \code{.x}. Filters \code{.x} if it is a list.}
}
\value{
An \code{eq_term} object.
}
\description{
Allows developers to edit term properties individually or in a filtered set
by applying a function to the current value.
}
\details{
There are several helper functions that save the developer from having to
enter the \code{.property} argument. They are documented here, but note that the
\code{.property} argument is not necessary for them (e.g.
\link{apply_estimate}).

When \code{.x} is a list of \code{eq_term} objects, the \code{.boolean_mask} argument will
filter the list and apply the changes to a subset of the list items in \code{.x}.
Furthermore, the filter will not change the order of the items in the list.
}
\section{Functions}{
\itemize{
\item \code{apply_estimate}: Apply function to the \code{estimate} property.
}}

\examples{
my_term <- eq_term(
    term = "sma", estimate = 0.25, variable = "\\\\phi",
    subscript = 2, superscript = "(i)"
)

apply_property(
    .x = my_term,
    .property = 'estimate',
    .f = function(x) { x * -1 }
)

}
\seealso{
Other eq_term modifiers: 
\code{\link{apply_variable_property}()},
\code{\link{eq_term}()},
\code{\link{set_property}()},
\code{\link{set_variable_property}()}
}
\concept{eq_term modifiers}
\keyword{internal}
