% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrests.R
\docType{data}
\name{arrests}
\alias{arrests}
\title{Arrest data from Gelman & Hill}
\format{
A tibble with 225 rows and 4 variables:
\describe{
\item{precinct}{An integer denoting the precinct identification number.}
\item{eth}{A factor with the coded race/ethnicity}
\item{stops}{The number of police stops}
\item{arrests}{The number of prior arrests (this is used as an offset variable in the book)}
}
}
\usage{
arrests
}
\description{
Arrest data from Gelman & Hill's book, used in Chapter 6 (and others).
The data have been aggregated by precinct and race/ethnicity, with the
sum of prior arrests and stops calculated. You can download the original
data here: http://www.stat.columbia.edu/~gelman/arm/examples/police/
}
\keyword{datasets}
