% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{build_tex}
\alias{build_tex}
\title{Build a complete TeX equation}
\usage{
build_tex(lhs, rhs, ital_vars = ital_vars, use_coefs = use_coefs)
}
\arguments{
\item{lhs}{Left-hand side of the equation; character; comes from
\code{extract_lhs}}

\item{rhs}{Right-hand side of the equation; list with nested elements; comes
from \code{extract_rhs}}

\item{ital_vars}{Passed from \code{extract_eq}}

\item{use_coefs}{Passed from \code{extract_eq}}
}
\value{
A character string
}
\description{
Combine the left-hand and right-hand sides of a model, wrap them with TeX
commands, and combine them into a single equation
}
\keyword{internal}
