% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_term.R
\name{c_term}
\alias{c_term}
\title{Constructor for a custom term object.}
\usage{
c_term(...)
}
\arguments{
\item{...}{One or more R objects convertible to a character.}
}
\value{
A list classed as \code{eq_term}.
}
\description{
Returns a list of class \code{c_term} which parses similar to \code{eq_term}.
}
\details{
When a utility function hasn't been built or would be too niche, developers
can use a character vector (including \code{eq_term} objects) to build a term-like
object. \code{eq_term} and term-like objects will use standard
\code{convert2latex()} options when parsing. Ohters will be converted to a
character.
}
\examples{
c_term("\\\\sim \\\\mathcal{N}(", my_term, "\\\\,\\\\sigma^{2})")

}
\seealso{
\link{eq_term} for a full \code{eq_term} object constructor.

Other term constructors: 
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()}

Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}
}
\concept{c_term constructors}
\concept{extensibility functions}
\concept{term constructors}
\keyword{internal}
