% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_by.R
\name{combine_by}
\alias{combine_by}
\alias{combine_by.default}
\alias{combine_by.character}
\alias{combine_by.function}
\title{Combine terms.}
\usage{
combine_by(.x, .combiner)

\method{combine_by}{default}(.x, .combiner)

\method{combine_by}{character}(.x, .combiner)

\method{combine_by}{`function`}(.x, .combiner)
}
\arguments{
\item{.x}{One or more \code{eq_term} or term-like objects.}

\item{.combiner}{A character or function. Used for dispatch and combining the
terms. These may have different behavior. Please check the documentation
for individual methods. \emph{Functions must return a \code{c_term} class!}}
}
\value{
A list.
}
\description{
Allows developers to combine a set of terms (i.e. by a operator like
\code{"+"} or a function).
}
\details{
The function dispatches on the \code{.combiner} argument. For example, a
character will dispatch to \code{combine_by.character} and will place the
character between the operators.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for \code{combine_by()}. Returns an
error.

\item \code{character}: Combines terms using a character operator between them (e.g. \code{"+"})

\item \code{function}: Combines terms using the provided function, returning a list of \code{c_term} objects.
}}

\seealso{
When nested in another list, use \code{flatten_nested} to bring up one
level.

Other term combiners: 
\code{\link{eq_divide}()}

Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}
}
\concept{extensibility functions}
\concept{term combiners}
\keyword{internal}
