% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_divide.R
\name{convert2latex.eq_divide}
\alias{convert2latex.eq_divide}
\alias{as.character.eq_divide}
\title{Convert \code{eq_divide} object to Latex}
\usage{
\method{convert2latex}{eq_divide}(
  .x,
  .use_coef = FALSE,
  .coef_digits = 2,
  .use_coef_sign = TRUE,
  .inverse_sign = FALSE,
  ...
)

\method{as.character}{eq_divide}(.x, ...)
}
\arguments{
\item{.x}{A convertible object (e.g. an \code{eq_term} object).}

\item{.use_coef}{A logical. Switch usage of the \code{estimate} or \code{variable}
representation. Default: FALSE.}

\item{.coef_digits}{A numeric used to round the term \code{estimate}. Default: 2.}

\item{.use_coef_sign}{A logical. If FALSE, will treat all signs on \code{estimate} as positive. Default: TRUE.}

\item{.inverse_sign}{A logical. If TRUE, will inverse the sign on \code{estimate}. Default: FALSE.}

\item{...}{Other parameters that may need passed to downstream functions.}
}
\value{
A latex-friendly character (string) representation of the term.
}
\description{
Converts a \code{eq_divide} classed object latex-friendly character string.
}
\details{
Since class-specific methods of this generic may have a wide variety of
arguments, these methods will have separate documentation pages. For example,
\link{convert2latex.eq_term()} is a separate document Use
\verb{?convert2latex.<class name>} or the \emph{See Also} section below.
}
\examples{
my_term <- eq_term(
    term = "sma", estimate = 0.25, variable = "\\\\phi",
    subscript = 2, superscript = "(i)"
)

convert2latex(.x = my_term)

}
\seealso{
Other convert2latex functions: 
\code{\link{convert2latex.c_term}()},
\code{\link{convert2latex.eq_equation}()},
\code{\link{convert2latex.eq_line}()},
\code{\link{convert2latex.eq_term}()},
\code{\link{convert2latex.eq_variable}()},
\code{\link{convert2latex}()}
}
\concept{convert2latex functions}
\keyword{internal}
