% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_line.R
\name{convert2latex.eq_line}
\alias{convert2latex.eq_line}
\alias{as.character.eq_line}
\title{Convert \code{eq_line} object to Latex}
\usage{
\method{convert2latex}{eq_line}(
  .x,
  .use_coef = FALSE,
  .coef_digits = 2,
  .use_coef_sign = TRUE,
  .inverse_sign = FALSE,
  .wrap = FALSE,
  .operator_location = "end",
  .terms_per_line = 4,
  ...
)

\method{as.character}{eq_line}(.x, ...)
}
\arguments{
\item{.x}{A convertible object (e.g. an \code{eq_term} object).}

\item{.use_coef}{A logical. Switch usage of the \code{estimate} or \code{variable}
representation. Default: FALSE.}

\item{.coef_digits}{A numeric used to round the term \code{estimate}. Default: 2.}

\item{.use_coef_sign}{A logical. If FALSE, will treat all signs on \code{estimate} as positive. Default: TRUE.}

\item{.inverse_sign}{A logical. If TRUE, will inverse the sign on \code{estimate}. Default: FALSE.}

\item{.wrap}{A logical. Whether or not to wrap terms in the line. Default is
disabled (FALSE).}

\item{.operator_location}{One of \code{c("start", "end")}. When terms are
split across multiple lines, they are split at mathematical operators like
\code{+}.When \dQuote{start}, each line will begin with an operator. When
\dQuote{end} each line will end with a trailing operator. Default is
\dQuote{end}.}

\item{.terms_per_line}{An integer. How many terms in a line before wrapping,
if wrapping is enabled. Will be rounded down if not an integer. Default is 4.}

\item{...}{Other parameters that may need passed to downstream functions.}
}
\value{
A latex-friendly character (string) representation of the term.
}
\description{
Converts a \code{eq_line} classed object latex-friendly character string.
}
\details{
Since class-specific methods of this generic may have a wide variety of
arguments, these methods will have separate documentation pages. For example,
\link{convert2latex.eq_term()} is a separate document Use
\verb{?convert2latex.<class name>} or the \emph{See Also} section below.
}
\examples{
my_term <- eq_term(
    term = "sma", estimate = 0.25, variable = "\\\\phi",
    subscript = 2, superscript = "(i)"
)

convert2latex(.x = my_term)

}
\seealso{
Other convert2latex functions: 
\code{\link{convert2latex.c_term}()},
\code{\link{convert2latex.eq_divide}()},
\code{\link{convert2latex.eq_equation}()},
\code{\link{convert2latex.eq_term}()},
\code{\link{convert2latex.eq_variable}()},
\code{\link{convert2latex}()}
}
\concept{convert2latex functions}
\keyword{internal}
