% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_variable.R
\name{convert2latex.eq_variable}
\alias{convert2latex.eq_variable}
\alias{as.character.eq_variable}
\title{Convert \code{eq_variable} object to Latex}
\usage{
\method{convert2latex}{eq_variable}(.x, ...)

\method{as.character}{eq_variable}(.x, ...)
}
\arguments{
\item{.x}{A convertible object (e.g. an \code{eq_term} object).}

\item{...}{Other parameters that may need passed to downstream functions.}
}
\value{
A latex-friendly character (string) representation of the variable.
}
\description{
Converts a \code{eq_variable} classed object latex-friendly character string.
}
\details{
Since class-specific methods of this generic may have a wide variety of
arguments, these methods will have separate documentation pages. For example,
\link{convert2latex.eq_term()} is a separate document Use
\verb{?convert2latex.<class name>} or the \emph{See Also} section below.
}
\examples{
my_term <- eq_term(
    term = "sma", estimate = 0.25, variable = "\\\\phi",
    subscript = 2, superscript = "(i)"
)

convert2latex(.x = my_term)

}
\seealso{
Other convert2latex functions: 
\code{\link{convert2latex.c_term}()},
\code{\link{convert2latex.eq_divide}()},
\code{\link{convert2latex.eq_equation}()},
\code{\link{convert2latex.eq_line}()},
\code{\link{convert2latex.eq_term}()},
\code{\link{convert2latex}()}
}
\concept{convert2latex functions}
\keyword{internal}
