% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{convert2latex}
\alias{convert2latex}
\alias{convert2latex.default}
\title{Convert to Latex}
\usage{
convert2latex(.x, ...)

\method{convert2latex}{default}(.x, ...)
}
\arguments{
\item{.x}{A convertible object (e.g. an \code{eq_term} object).}

\item{...}{Other arguments that may be used by downstream functions.}
}
\value{
A latex-friendly character (string) representation of the term.
}
\description{
Converts a compatible classed object or subclass to a latex-friendly
character string.
}
\details{
Since class-specific methods of this generic may have a wide variety of
arguments, these methods will have separate documentation pages. For example,
\link{convert2latex.eq_term()} is a separate document Use
\verb{?convert2latex.<class name>} or the \emph{See Also} section below.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Returns as.character(.x) and serves a warning.
}}

\examples{
my_term <- eq_term(
    term = "sma", estimate = 0.25, variable = "\\\\phi",
    subscript = 2, superscript = "(i)"
)

convert2latex(.x = my_term)

}
\seealso{
Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}

Other convert2latex functions: 
\code{\link{convert2latex.c_term}()},
\code{\link{convert2latex.eq_divide}()},
\code{\link{convert2latex.eq_equation}()},
\code{\link{convert2latex.eq_line}()},
\code{\link{convert2latex.eq_term}()},
\code{\link{convert2latex.eq_variable}()}
}
\concept{convert2latex functions}
\concept{extensibility functions}
\keyword{internal}
