% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{convert2terms}
\alias{convert2terms}
\alias{convert2terms.default}
\alias{convert2terms.data.frame}
\alias{as.eq_term}
\title{Convert to \code{eq_term}}
\usage{
convert2terms(.x, .include, ...)

\method{convert2terms}{default}(.x, ...)

\method{convert2terms}{data.frame}(.x, .include, ...)

as.eq_term(.x, .include, ...)
}
\arguments{
\item{.x}{A convertible object (e.g. a \code{\link[=broom:tidy]{broom:tidy()}} \code{tbl_df}).}

\item{.include}{A character vector of additional columns (etc.) to include as
properties of the \code{eq_term} class. If missing, will include all columns
(etc.) as properties.}

\item{...}{Other arguments that may be used by downstream functions.}
}
\value{
A \code{eq_term} classed list or list of \code{eq_term} classed lists.
}
\description{
Converts a compatible classed object or subclass to an \code{eq_term} or list of
\code{eq_term}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Throws an error if a method isn't found.

\item \code{data.frame}: Converts \code{data.frame} to a list of \code{eq_term}.
}}

\examples{
df <- data.frame(
    term = c("ar1", "ma1", "drift"),
    estimate = c(0.02, 0.1, 0.3)
)

convert2terms(df)

}
\seealso{
\link{as.eq_term} wraps this function.

Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}
}
\concept{convert2terms functions}
\concept{extensibility functions}
\keyword{internal}
