% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{create_eq.default}
\alias{create_eq.default}
\title{Create the full equation}
\usage{
\method{create_eq}{default}(
  lhs,
  rhs,
  ital_vars,
  use_coefs,
  coef_digits,
  fix_signs,
  model,
  intercept,
  greek,
  raw_tex
)
}
\arguments{
\item{lhs}{A character string of the left-hand side variable extracted with
\code{extract_lhs}}

\item{rhs}{A data frame of right-hand side variables extracted with
\code{extract_rhs}.}

\item{ital_vars}{Logical, defaults to \code{FALSE}. Should the variable names
not be wrapped in the \code{\\operatorname{}} command?}

\item{use_coefs}{Logical, defaults to \code{FALSE}. Should the actual model
estimates be included in the equation instead of math symbols?}

\item{coef_digits}{Integer, defaults to 2. The number of decimal places to
round to when displaying model estimates.}

\item{fix_signs}{Logical, defaults to \code{FALSE}. If disabled,
coefficient estimates that are negative are preceded with a "+" (e.g.
\code{5(x) + -3(z)}). If enabled, the "+ -" is replaced with a "-" (e.g.
\code{5(x) - 3(z)}).}

\item{model}{A fitted model}

\item{intercept}{How should the intercept be displayed? Default is \code{"alpha"},
but can also accept \code{"beta"}, in which case the it will be displayed
as beta zero.}

\item{greek}{What notation should be used for
coefficients? Currently only accepts \code{"beta"} (with plans for future
development). Can be used in combination with \code{raw_tex} to use any
notation, e.g., \code{"\\hat{\\beta}"}.}

\item{raw_tex}{Logical. Is the greek code being passed to denote coefficients
raw tex code?}
}
\description{
Create the full equation
}
\keyword{internal}
