% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{create_term}
\alias{create_term}
\title{Create a full term w/subscripts}
\usage{
create_term(rhs, ital_vars)
}
\arguments{
\item{rhs}{A data frame of right-hand side variables extracted with
\code{extract_rhs}.}

\item{ital_vars}{Logical, defaults to \code{FALSE}. Should the variable names
not be wrapped in the \code{\\operatorname{}} command?}
}
\description{
Create a full term w/subscripts
}
\keyword{internal}
