% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_divide.R
\name{eq_divide}
\alias{eq_divide}
\title{Divide \code{eq_term} objects in Latex equation.}
\usage{
eq_divide(.numerator, .denominator)
}
\arguments{
\item{.numerator}{A character or \code{eq_term} object.}

\item{.denominator}{A character or \code{eq_term} object.}
}
\value{
A list classed as \code{eq_division}.
}
\description{
Returns a list of class \code{eq_divide} which parses similar to \code{eq_term}.
}
\details{
One of several helper functions to provide for consistent parsing to Latex.
}
\examples{
eq_divide(myeqterm, "5")

}
\seealso{
Other term combiners: 
\code{\link{combine_by}()}

Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}
}
\concept{extensibility functions}
\concept{term combiners}
\keyword{internal}
