% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_equation.R
\name{eq_equation}
\alias{eq_equation}
\title{Build an \code{equatiomatic} equation.}
\usage{
eq_equation(
  ...,
  .wrapable_operators = c("=", "+", "-"),
  .special_class = NA_character_
)
}
\arguments{
\item{...}{One or more R objects convertible to a character, including \code{eq_line}.}

\item{.wrapable_operators}{A character vector. Represents the character
objects where the equation can be wrapped. Default is \code{c("=", "+", "-")}.}

\item{.special_class}{A character. A class, preferably the model class
prefixed with \dQuote{eq_}, that will be used for dispatch. Useful when
special wrapping or alignment functions are required.}
}
\value{
A list classed as \code{eq_equation} or a special class as provided.
}
\description{
Final equation generated by \code{equatiomatic} from a model.
}
\details{
The final equation object produced by \code{equatiomatic} from a model.
}
\examples{
eq_equation(c_term("y0"), "=", myterm1, "+", myterm2)

}
\seealso{
Other term constructors: 
\code{\link{c_term}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()}

Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}
}
\concept{creating equations}
\concept{extensibility functions}
\concept{term constructors}
\keyword{internal}
