% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_line.R
\name{eq_line}
\alias{eq_line}
\title{Encapsulate a line in an equation.}
\usage{
eq_line(..., .wrapable_operators = c("=", "+", "-"))
}
\arguments{
\item{...}{One or more R objects convertible to a character.}

\item{.wrapable_operators}{A character vector. Represents the character
objects where the line can be wrapped. Default is \code{c("=", "+", "-")}.}
}
\value{
A list classed as \code{eq_line}.
}
\description{
An encapsulated line that wraps separate of other lines in an equation.
}
\details{
One of several helper functions to provide for consistent parsing to Latex.
\code{eq_line} allows developers to encapsulate terms that should remain on their
own line and wrap together.

Null objects provided in \code{...} will be removed along with the preceding
element.
}
\examples{
eq_line(c_term("y0"), "=", myterm1, "+", myterm2)

}
\seealso{
Other term constructors: 
\code{\link{c_term}()},
\code{\link{eq_equation}()},
\code{\link{eq_term}()}

Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()},
\code{\link{wrap_eq}()}
}
\concept{extensibility functions}
\concept{term constructors}
\keyword{internal}
