% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equatiomatic-package.R
\docType{package}
\name{equatiomatic-package}
\alias{equatiomatic}
\alias{equatiomatic-package}
\title{equatiomatic: Transform Models into LaTeX Equations}
\description{
The goal of equatiomatic is to reduce the pain associated with
             writing LaTex formulas from fitted models. The primary function of
             the package, extract_eq_lm, takes a fitted lm object as its input
             and returns the corresponding LaTeX code for the model.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (https://orcid.org/0000-0003-4699-4680)

Authors:
\itemize{
  \item Joshua Rosenberg \email{jmrosenberg@utk.edu}
  \item Andrew Heiss \email{andrew@andrewheiss.com} (https://orcid.org/0000-0002-3948-3914)
}

Other contributors:
\itemize{
  \item Jonathan Sidi \email{yonis@metrumrg.com} (https://orcid.org/0000-0002-4222-1819) [contributor]
}

}
\keyword{internal}
