% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{escape_tex}
\alias{escape_tex}
\title{Escape TeX}
\usage{
escape_tex(term)
}
\arguments{
\item{term}{A character string to escape}
}
\value{
A character string
}
\description{
Escape special TeX characters.
}
\details{
Ten characters have special meaning in TeX \code{& \% $ # _ { } ~ ^ \\}.
This function either escapes them with \\, or in the case of the last three,
replaces them with special TeX macros.
}
\keyword{internal}
