% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rhs.R
\name{extract_all_subscripts}
\alias{extract_all_subscripts}
\title{Extract all subscripts}
\usage{
extract_all_subscripts(primary_list, full_term_list)
}
\arguments{
\item{primary_list}{A list of primary terms}

\item{full_term_list}{A list of full terms}
}
\value{
A list with the subscripts. If full term has no subscript,
returns \code{""}.
}
\description{
Extract all subscripts
}
\examples{
\dontrun{
p_list <- list("partyid",
               c("partyid", "age"),
               c("age", "race"),
               c("partyid", "age", "race"))

ft_list <- list("partyidNot str republican",
                c("partyidInd,near dem", "age"),
                c("age", "raceBlack"),
                c("partyidInd,near dem", "age", "raceBlack"))

extract_all_subscripts(p_list, ft_list)
}
}
\keyword{internal}
