% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ARIMA.R
\name{extract_eq.forecast_ARIMA}
\alias{extract_eq.forecast_ARIMA}
\title{'LaTeX' code for forecast_ARIMA models.}
\usage{
\method{extract_eq}{forecast_ARIMA}(
  .model,
  .ital_vars = FALSE,
  .wrap = FALSE,
  .terms_per_line = 4,
  .operator_location = "end",
  .align_env = "aligned",
  .use_coefs = FALSE,
  .coef_digits = 2,
  .fix_signs = TRUE,
  .font_size
)
}
\arguments{
\item{.model}{A fitted model.}

\item{.ital_vars}{A logical. Setting to \code{TRUE} will use standard latex
parsing for text (italicized). Default is \code{FALSE}.}

\item{.wrap}{A logical. When set to \code{TRUE}, will wrap equations where
possible. Default is \code{FALSE}.}

\item{.terms_per_line}{An integer. The number of allowable terms per line.
Used only when \code{.wrap = TRUE}. Default is \code{4}.}

\item{.operator_location}{One of \code{c("start", "end")}. Determines if a
wrapped operator will end a line or start on the next one. Default is
\dQuote{end}.}

\item{.align_env}{One of \code{c("aligned", "aligned*", "align", "align*")}.
TeX environment to wrap around equation. Default is \dQuote{aligned}.}

\item{.use_coefs}{A logical. Toggle use of model coefficients (estimates) or
mathematical representations. Default is \code{FALSE}.}

\item{.coef_digits}{An integer. The number of decimal places to round to when
displaying model coefficients (estimates). Default is \code{2}.}

\item{.fix_signs}{A logical. If \code{FALSE}, coefficient estimates that are
negative are preceded with a "+" (e.g. \code{5(x) + -3(z)}). If \code{TRUE}, the
"+ -" is replaced with a "-" (e.g. \code{5(x) - 3(z)}). Default is \code{FALSE}.}

\item{.font_size}{The font size of the equation. Defaults to default of
the output format. Takes any of the standard LaTeX arguments (see
\href{https://www.overleaf.com/learn/latex/Font_sizes,_families,_and_styles#Font_styles}{here}).}
}
\value{
A character of class \dQuote{equation}.
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
}
\details{
Extracts equation from \code{forecast::Arima()} built model.

Implementations for different models may have additional arguments. Please
see model-specific documentation for a description of those arguments.
}
\seealso{
Other extraction functions: 
\code{\link{extract_eq}()}
}
\concept{extraction functions}
