% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_lm.R
\name{extract_eq_lm}
\alias{extract_eq_lm}
\title{Latex code for `lm` models}
\usage{
extract_eq_lm(model, preview = FALSE, use_text = FALSE)
}
\arguments{
\item{model}{A fitted `lm` model}

\item{preview}{Logical, defaults to \code{FALSE}. Should the equation be
previewed in the viewer pane?}

\item{use_text}{Logical, defaults to \code{FALSE}. Should the variable names
be wrapped in the \code{\\text{}} command}
}
\description{
Uses the variable names supplied to \link[stats]{lm} to produce a LaTeX
equation, which is output to the screen.
}
\examples{

# Simple model
mod1 <- lm(mpg ~ cyl + disp, mtcars)
extract_eq_lm(mod1)

# Include all variables
mod2 <- lm(mpg ~ ., mtcars)
extract_eq_lm(mod2)

# Works for categorical variables too, putting levels as subscripts
mod3 <- lm(Sepal.Length ~ Sepal.Width + Species, iris)
extract_eq_lm(mod3)

set.seed(8675309)
d <- data.frame(cat1 = rep(letters[1:3], 100),
                cat2 = rep(LETTERS[1:3], each = 100),
                cont1 = rnorm(300, 100, 1),
                cont2 = rnorm(300, 50, 5),
                out   = rnorm(300, 10, 0.5))
mod4 <- lm(out ~., d)
extract_eq_lm(mod4)

# Or preview the equation
extract_eq_lm(mod4, preview = TRUE)

# Use \\text{}
# Simple model
extract_eq_lm(mod1, use_text = TRUE)

# Categorical variables
extract_eq_lm(mod3, use_text = TRUE)
}
