% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lhs.R
\name{extract_lhs.clm}
\alias{extract_lhs.clm}
\title{Extract left-hand side of a clm object}
\usage{
\method{extract_lhs}{clm}(model, ital_vars, ...)
}
\arguments{
\item{model}{A fitted model}

\item{ital_vars}{Logical, defaults to \code{FALSE}. Should the variable names
not be wrapped in the \code{\\operatorname{}} command?}
}
\value{
A character string
}
\description{
Extract a string of the outcome/dependent/y variable with the appropriate
link function.
}
\keyword{internal}
