% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lhs.R
\name{extract_lhs.lm}
\alias{extract_lhs.lm}
\title{Extract left-hand side of an lm object}
\usage{
\method{extract_lhs}{lm}(model, ital_vars, ...)
}
\arguments{
\item{model}{A fitted model}

\item{ital_vars}{Logical, defaults to \code{FALSE}. Should the variable names
not be wrapped in the \code{\\operatorname{}} command?}
}
\value{
A character string
}
\description{
Extract a string of the outcome/dependent/y variable of a model
}
\keyword{internal}
