% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{extract_lhs}
\alias{extract_lhs}
\title{Extract left-hand side}
\usage{
extract_lhs(model, ital_vars)
}
\arguments{
\item{model}{A fitted model}
}
\value{
A character string
}
\description{
Extract a string of the outcome/dependent/y variable of a model
}
\keyword{internal}
