% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rhs.R
\name{extract_primary_term}
\alias{extract_primary_term}
\title{Extract the primary terms from all terms}
\usage{
extract_primary_term(primary_term_v, all_terms)
}
\arguments{
\item{primary_term_v}{A vector of primary terms, e.g., \code{"partyid"}.
Usually the result of \code{formula_rhs[!grepl(":", formula_rhs)]}}

\item{all_terms}{A list of all the equation terms on the right hand side,
usually the result of \code{broom::tidy(model, quick = TRUE)$term}.}
}
\description{
Extract the primary terms from all terms
}
\examples{
\dontrun{
primaries <- c("partyid", "age", "race")

full_terms <- c("partyidDon't know", "partyidOther party", "age",
"partyidNot str democrat", "age", "raceBlack", "age", "raceBlack")

extract_primary_term(primaries, full_terms)
}
}
\keyword{internal}
