% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{extract_rhs}
\alias{extract_rhs}
\title{Extract right-hand side}
\usage{
extract_rhs(model)
}
\arguments{
\item{model}{A fitted model}
}
\value{
A list with one element per future equation term. Term components
  like subscripts are nested inside each list element. List elements with two
  or more terms are interactions.
}
\description{
Extract a nested list of the explanatory/independent/x variables of a model
}
\examples{

mod1 <- lm(Sepal.Length ~ Sepal.Width + Species * Petal.Length, iris)

str(extract_rhs(mod1))
#> List of 6
#>  $ Sepal.Width                   :List of 1
#>   ..$ :List of 1
#>   .. ..$ term: chr "Sepal.Width"
#>  $ Speciesversicolor             :List of 1
#>   ..$ :List of 2
#>   .. ..$ term     : chr "Species"
#>   .. ..$ subscript: chr "versicolor"
#>  $ Speciesvirginica              :List of 1
#>   ..$ :List of 2
#>   .. ..$ term     : chr "Species"
#>   .. ..$ subscript: chr "virginica"
#>  $ Petal.Length                  :List of 1
#>   ..$ :List of 1
#>   .. ..$ term: chr "Petal.Length"
#>  $ Speciesversicolor:Petal.Length:List of 2
#>   ..$ :List of 2
#>   .. ..$ term     : chr "Species"
#>   .. ..$ subscript: chr "versicolor"
#>   ..$ :List of 1
#>   .. ..$ term: chr "Petal.Length"
#>  $ Speciesvirginica:Petal.Length :List of 2
#>   ..$ :List of 2
#>   .. ..$ term     : chr "Species"
#>   .. ..$ subscript: chr "virginica"
#>   ..$ :List of 1
#>   .. ..$ term: chr "Petal.Length"

}
\keyword{internal}
