% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rhs.R
\name{extract_subscripts}
\alias{extract_subscripts}
\title{Extract the subscripts from a given term}
\usage{
extract_subscripts(primary, full_term_v)
}
\arguments{
\item{primary}{A single primary term, e.g., \code{"partyid"}}

\item{full_term_v}{A vector of full terms, e.g.,
\code{c("partyidDon't know", "partyidOther party"}. Can be of length 1.}
}
\description{
Extract the subscripts from a given term
}
\examples{
\dontrun{
extract_subscripts("partyid", "partyidDon't know")
extract_subscripts("partyid",
                   c("partyidDon't know", "partyidOther party",
                     "partyidNot str democrat"))
}
}
\keyword{internal}
