% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{fix_coef_signs_se}
\alias{fix_coef_signs_se}
\title{Fixes the Signs When Using se_subscripts Argument}
\usage{
fix_coef_signs_se(equation)
}
\arguments{
\item{equation}{list that contains the equation}
}
\value{
a list containing the equation with fixed signs
}
\description{
Fixes the Signs When Using se_subscripts Argument
}
