% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_nested.R
\name{flatten_nested}
\alias{flatten_nested}
\title{Flatten Nested List of Terms}
\usage{
flatten_nested(.x)
}
\arguments{
\item{.x}{A list.}
}
\value{
A list.
}
\description{
Flattens eligible objects one level up.
}
\details{
Will flatten elements of the provided list where the object has the
\code{attribute} \code{flatten} set to \code{TRUE}.
}
\seealso{
Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{wrap_eq}()}
}
\concept{extensibility functions}
\keyword{internal}
