% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.equation}
\alias{format.equation}
\title{format 'LaTeX' equations}
\usage{
\method{format}{equation}(x, ...)
}
\arguments{
\item{x}{'LaTeX' equation built with \code{\link{extract_eq}}}

\item{...}{not used}
}
\description{
format 'LaTeX' equations built with \code{\link{extract_eq}}.
}
