% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lhs.R
\name{modify_lhs_for_link}
\alias{modify_lhs_for_link}
\title{modifies lhs of equations that include a link function}
\usage{
modify_lhs_for_link(model, ...)
}
\description{
modifies lhs of equations that include a link function
}
\keyword{internal}
