% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_term.R
\name{new_eq_term}
\alias{new_eq_term}
\title{Base constructor for \code{eq_term} class}
\usage{
new_eq_term(
  .term = NA_character_,
  .estimate = NA_real_,
  .variable = new_eq_variable(),
  .suffix = new_eq_variable(),
  .custom_parsing = function(x) {     x }
)
}
\arguments{
\item{.term}{A character representing the term name.}

\item{.estimate}{A number. Swaps with 'variable'.}

\item{.variable}{A \code{eq_variable}. Swaps with 'estimate'.}

\item{.suffix}{A \code{eq_variable}.}
}
\value{
A list classed as \code{eq_term}.
}
\description{
Constructs an \code{eq_term} object with minimal checks.
}
\details{
As a base constructor for \code{eq_term} class, this function only performs type checks, should
only be used by developers, and when the you know the types are correct. For
a fully-fledged constructors with coercion and validation see \code{\link[=eq_term]{eq_term()}}.
}
\seealso{
\link{eq_term} for a full constructor (including coercion and
validation), \link{validate_eq_term} for a more complex validation

Other eq_term constructors: 
\code{\link{eq_term}()},
\code{\link{validate_eq_term}()}
}
\concept{eq_term constructors}
\keyword{internal}
