% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_variable.R
\name{new_eq_variable}
\alias{new_eq_variable}
\title{Base constructor for \code{eq_variable} class}
\usage{
new_eq_variable(
  .x = NA_character_,
  .subscript = NA_character_,
  .superscript = NA_character_,
  .italic = TRUE
)
}
\arguments{
\item{.x}{A character representing the variable name.}

\item{.subscript}{A character.}

\item{.superscript}{A character.}

\item{.italic}{A logical.}
}
\value{
A list classed as \code{eq_term}.
}
\description{
Constructs an \code{eq_variable} object with minimal checks.
}
\details{
As a base constructor for \code{eq_variable} class, this function only performs type checks, should
only be used by developers, and when the you know the types are correct. For
a fully-fledged constructors with coercion and validation see \code{\link[=eq_variable]{eq_variable()}}.
}
\seealso{
\link{eq_term} for a full constructor (including coercion and
validation), \link{validate_eq_term} for a more complex validation

Other eq_variable constructors: 
\code{\link{eq_variable}()},
\code{\link{validate_eq_variable}()}
}
\concept{eq_variable constructors}
\keyword{internal}
