% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview LaTeX equations}
\usage{
preview(eq, ...)
}
\arguments{
\item{eq}{LaTeX equation built with \code{\link{extract_eq}}}

\item{\dots}{additional arguments passed to \code{\link[texPreview]{tex_preview}}}
}
\description{
Preview LaTeX equations built with \code{\link{extract_eq}}.
}
\examples{
\dontrun{
# Basic preview
mod1 <- lm(mpg ~ cyl + disp, mtcars)
preview(extract_eq(mod1))

# Pass arguments to tex_preview()
preview(extract_eq(mod1), density = 600)

# Use with magrittr pipes
mod1 \%>\%
  preview(density = 600)
}
}
