% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.equation}
\alias{print.equation}
\title{Print LaTeX equations}
\usage{
\method{print}{equation}(eq)
}
\arguments{
\item{eq}{LaTeX equation built with \code{\link{extract_eq}}}
}
\description{
Print LaTeX equations built with \code{\link{extract_eq}}.
}
