% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{print.tex}
\alias{print.tex}
\alias{print}
\title{Print equation}
\usage{
\method{print}{tex}(x, ...)
}
\arguments{
\item{x}{LaTeX equation built with \code{build_tex}}

\item{\dots}{arguments that control equation wrap width and equation environment
see details}
}
\description{
Print method for tex class
}
\details{
arguments that can be passed in the \dots

width: Controls the width of the equation, if this is set the the equation is aligned

template: the type of environment that the equation is wrapped in, e.g. $, $$ , align, equation
}
