% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderEq}
\alias{renderEq}
\alias{eqOutput}
\title{Display equations in shiny apps}
\usage{
renderEq(expr, env = parent.frame(), quoted = FALSE, outputArgs = list())

eqOutput(outputId)
}
\arguments{
\item{expr}{An R expression, specifically a call to \code{\link[=extract_eq]{extract_eq()}}}

\item{env}{The environment}

\item{quoted}{Is the expresion quoted?}

\item{outputArgs}{list of output arguments}

\item{outputId}{The identifier of the output from the server. Should be
passed as a string.}
}
\value{
Render the equation in a suitable way for Shiny for \code{\link[=renderEq]{renderEq()}} in
an \code{\link[=eqOutput]{eqOutput()}} equation output element that can be included in a panel.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
These are a set of functions designed to help render equations in
Shiny applications (see the vignette about Shiny).
}
\section{Functions}{
\itemize{
\item \code{renderEq()}: Rendering function

\item \code{eqOutput()}: Output function

}}
