% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_property.R
\name{set_property}
\alias{set_property}
\alias{set_variable}
\alias{set_suffix}
\title{Add or set an \code{eq_term} object property.}
\usage{
set_property(.x, .property, .new, .boolean_mask)

set_variable(.x, .new, .boolean_mask)

set_suffix(.x, .new, .boolean_mask)
}
\arguments{
\item{.x}{An \code{eq_term} object or list of \code{eq_term} objects.}

\item{.property}{A character or vector of character. Name(s) of property(ies) to
be modified. If a vector, must be the same length as \code{.new}.}

\item{.new}{A numeric, character, or \code{eq_term} or vector of them. The
value(s) that is(are) going to replace the current value(s) in the
property(s). If a vector, must be the same length as \code{.property};}

\item{.boolean_mask}{Optional. A vector of logicals (booleans) the same
length at \code{.x}. Filters \code{.x} if it is a list.}
}
\value{
An \code{eq_term} object.
}
\description{
Allows developers to add or edit term properties individually or in a
filtered set.
}
\details{
There are several helper functions that save the developer from having to
enter the \code{.property} argument. These are documented here, but note that the
\code{.property} argument is not necessary for them (e.g.
\link{set_variable}).

When \code{.x} is a list of \code{eq_term} objects, the \code{.boolean_mask} argument will
filter the list and apply the changes to a subset of the list items in \code{.x}.
Furthermore, the filter will not change the order of the items in the list.

\bold{Note:} Since this function can be used to \emph{add a property} there
is no check that the property actually exists. It will be created if it
doesn't exist. Make sure to check your spelling.
}
\section{Functions}{
\itemize{
\item \code{set_variable}: Set the \code{variable} property.

\item \code{set_suffix}: Set the \code{suffix} property.
}}

\examples{
my_term <- eq_term(
    term = "sma", estimate = 0.25, variable = "\\\\phi",
    subscript = 2, superscript = "(i)"
)

set_property(
    .x = my_term,
    .property = 'variable',
    .new = '\\\\beta'
)

}
\seealso{
Other eq_term modifiers: 
\code{\link{apply_property}()},
\code{\link{apply_variable_property}()},
\code{\link{eq_term}()},
\code{\link{set_variable_property}()}
}
\concept{eq_term modifiers}
\keyword{internal}
