% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variable_property.R
\name{set_variable_property}
\alias{set_variable_property}
\alias{set_variable_name}
\alias{set_variable_subscript}
\alias{set_variable_superscript}
\alias{set_variable_italic}
\title{Add or set a property of an \code{eq_variable} held in an \code{eq_term} object .}
\usage{
set_variable_property(
  .x,
  .term_property,
  .variable_property,
  .new,
  .boolean_mask
)

set_variable_name(.x, .term_property, .new, .boolean_mask)

set_variable_subscript(.x, .term_property, .new, .boolean_mask)

set_variable_superscript(.x, .term_property, .new, .boolean_mask)

set_variable_italic(.x, .term_property, .new, .boolean_mask)
}
\arguments{
\item{.x}{An \code{eq_term} object or list of \code{eq_term} objects.}

\item{.term_property}{A character or vector of characters. Name(s) of the
properties in the \code{eq_term} object that hold the \code{eq_variable} objects to
be modified.}

\item{.variable_property}{A character or vector of character. Name(s) of
property(ies) in the \code{eq_variable} object to be modified. If a vector, must
be the same length as \code{.new}.}

\item{.new}{A numeric, character, or \code{eq_term} or vector of them. The
value(s) that is(are) going to replace the current value(s) in the
property(ies). If a vector, must be the same length as \code{.property};}

\item{.boolean_mask}{Optional. A vector of logicals (booleans) the same
length at \code{.x}. Filters \code{.x} if it is a list.}
}
\value{
An \code{eq_term} object.
}
\description{
Allows developers to add or edit term properties individually or in a
filtered set.
}
\details{
There are several helper functions that save the developer from having to
enter the \code{.variable_property} argument. These are documented here, but note that the
\code{.variable_property} argument is not necessary for them (e.g.
\link{set_variable}).

When \code{.x} is a list of \code{eq_term} objects, the \code{.boolean_mask} argument will
filter the list and apply the changes to a subset of the list items in \code{.x}.
Furthermore, the filter will not change the order of the items in the list.

\bold{Note:} Since this function can be used to \emph{add a property} there
is no check that the property actually exists. It will be created if it
doesn't exist. Make sure to check your spelling.
}
\section{Functions}{
\itemize{
\item \code{set_variable_name}: Set the \code{name} property.

\item \code{set_variable_subscript}: Set the \code{subscript} property.

\item \code{set_variable_superscript}: Set the \code{superscript} property.

\item \code{set_variable_italic}: Set the \code{italic} property.
}}

\examples{
my_term <- eq_term(term = "sma", estimate = 0.25)

set_variable_property(
    .x = my_term,
    .term_property = 'variable'
    .variable_property = 'name',
    .new = '\\\\beta'
)

}
\seealso{
Other eq_term modifiers: 
\code{\link{apply_property}()},
\code{\link{apply_variable_property}()},
\code{\link{eq_term}()},
\code{\link{set_property}()}
}
\concept{eq_term modifiers}
\keyword{internal}
