% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ts.R
\docType{data}
\name{simple_ts}
\alias{simple_ts}
\title{Simple simulated time series data}
\format{
A tibble with 1000 rows and 8 variables:
\describe{
\item{Qtr1}{First quarter simulated values.}
\item{Qtr2}{Second quarter simulated values.}
\item{Qtr3}{Third quarter simulated values.}
\item{Qtr4}{Fourth quarter simulated values.}
}
}
\usage{
simple_ts
}
\description{
Output from \code{set.seed(123); simple_ts <- ts(rnorm(1000),freq = 4)}.
This is included primarily for unit testing.
}
\keyword{datasets}
