% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{texify_term}
\alias{texify_term}
\title{TeXify an equation term}
\usage{
texify_term(term, ital_vars)
}
\arguments{
\item{term}{A character string to TeXify}

\item{ital_vars}{Passed from \code{extract_eq}}
}
\value{
A character string
}
\description{
Prepare an equation term for TeX, wrapping the text with \code{\\text{}}
if necessary
}
\keyword{internal}
