% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_term.R
\name{validate_eq_term}
\alias{validate_eq_term}
\title{Validation function for \code{eq_term} class}
\usage{
validate_eq_term(.eq_term)
}
\arguments{
\item{.eq_term}{An unvalidated \code{eq_term}.}
}
\value{
A validated \code{eq_term}.
}
\description{
Performs more complex checks when \code{eq_term} is constructed with the full constructor.
}
\seealso{
\link{eq_term} for a full constructor (including coercion and
validation), \link{new_eq_term} for a low-level constructor

Other eq_term constructors: 
\code{\link{eq_term}()},
\code{\link{new_eq_term}()}
}
\concept{eq_term constructors}
\keyword{internal}
