% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_variable.R
\name{validate_eq_variable}
\alias{validate_eq_variable}
\title{Validation function for \code{eq_variable} class}
\usage{
validate_eq_variable(.eq_variable)
}
\arguments{
\item{.eq_variable}{An unvalidated \code{eq_variable}.}
}
\value{
A validated \code{eq_variable}.
}
\description{
Performs more complex checks when \code{eq_variable} is constructed with the full constructor.
}
\seealso{
\link{eq_variable} for a full constructor (including coercion and
validation), \link{new_eq_variable} for a low-level constructor

Other eq_variable constructors: 
\code{\link{eq_variable}()},
\code{\link{new_eq_variable}()}
}
\concept{eq_variable constructors}
\keyword{internal}
