% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{wrap_eq}
\alias{wrap_eq}
\title{Wraps terms.}
\usage{
wrap_eq(.x, .chrs, .terms_per_line, .operator_location, ...)
}
\arguments{
\item{.x}{An equatiomatic term object (e.g. eq_equation or eq_line).}

\item{.chrs}{A list of characters. List of converted terms that will be modified and returned.}

\item{.terms_per_line}{An integer. The number of allowable terms per line.
Used only when \code{.wrap = TRUE}. Default is \code{4}.}

\item{.operator_location}{One of \code{c("start", "end")}. Determines if a
wrapped operator will end a line or start on the next one. Default is
\dQuote{end}.}

\item{...}{Other arguments that may be used by downstream functions.}
}
\value{
A \code{eq_term} classed list or list of \code{eq_term} classed lists.
}
\description{
Consolidation of wrapping logic.
}
\seealso{
Other extensibility functions: 
\code{\link{align_eq}()},
\code{\link{c_term}()},
\code{\link{combine_by}()},
\code{\link{convert2latex}()},
\code{\link{convert2terms}()},
\code{\link{eq_divide}()},
\code{\link{eq_equation}()},
\code{\link{eq_line}()},
\code{\link{eq_term}()},
\code{\link{eq_variable}()},
\code{\link{flatten_nested}()}

Other equatiomatic utility functions: 
\code{\link{align_eq}()}
}
\concept{equatiomatic utility functions}
\concept{extensibility functions}
\keyword{internal}
