% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rhs.R
\name{wrap_rhs}
\alias{wrap_rhs}
\title{Generic function for wrapping the RHS of a model equation in something, like
how the RHS of probit is wrapped in φ()}
\usage{
wrap_rhs(model, tex, ...)
}
\arguments{
\item{model}{A fitted model}

\item{tex}{The TeX version of the RHS of the model (as character), built as
\code{rhs_combined} or \code{eq_raw$rhs} in \code{extract_eq()}}

\item{\dots}{additional arguments passed to the specific extractor}
}
\description{
Generic function for wrapping the RHS of a model equation in something, like
how the RHS of probit is wrapped in φ()
}
\keyword{internal}
