% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_tex_ital}
\alias{add_tex_ital}
\title{Wrap text in \code{\\operatorname{}}}
\usage{
add_tex_ital(term, ital_vars)
}
\arguments{
\item{term}{A character to wrap in \code{\\operatorname{}}}

\item{ital_vars}{Passed from \code{extract_eq}}
}
\value{
A character string
}
\description{
Add tex code to make string not italicized within an equation
}
\keyword{internal}
