% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_tex_ital_v}
\alias{add_tex_ital_v}
\title{Wrap text in \code{\\operatorname{}} (vectorized)}
\usage{
add_tex_ital_v(term_v, ital_vars)
}
\value{
A vector of characters
}
\description{
Add tex code to make string not italicized within an equation for a vector
of strings
}
\keyword{internal}
