% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_tex_subscripts}
\alias{add_tex_subscripts}
\title{Wrap text in \code{_{}}}
\usage{
add_tex_subscripts(term)
}
\arguments{
\item{term}{A character string to TeXify}
}
\value{
A character string
}
\description{
Add tex code to make subscripts for a single string
}
\keyword{internal}
