% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{add_tex_subscripts_v}
\alias{add_tex_subscripts_v}
\title{Wrap text in \code{_{}}}
\usage{
add_tex_subscripts_v(term_v)
}
\value{
A vector of characters
}
\description{
Add tex code to make subscripts for a vector of strings
}
\keyword{internal}
