% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equatiomatic-package.R
\docType{package}
\name{equatiomatic-package}
\alias{equatiomatic}
\alias{equatiomatic-package}
\title{equatiomatic: Transform Models into 'LaTeX' Equations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'equatiomatic' is to reduce the pain associated with writing 'LaTeX' formulas from fitted models. The primary function of the package, extract_eq(), takes a fitted model object as its input and returns the corresponding 'LaTeX' code for the model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datalorax/equatiomatic}
  \item \url{https://datalorax.github.io/equatiomatic/}
  \item Report bugs at \url{https://github.com/datalorax/equatiomatic/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})
  \item Andrew Heiss \email{andrew@andrewheiss.com} (\href{https://orcid.org/0000-0002-3948-3914}{ORCID})
  \item Jay Sumners \email{jay.sumners@gmail.com}
}

Other contributors:
\itemize{
  \item Joshua Rosenberg \email{jmrosenberg@utk.edu} (\href{https://orcid.org/0000-0003-2170-0447}{ORCID}) [contributor]
  \item Jonathan Sidi \email{yonis@metrumrg.com} (\href{https://orcid.org/0000-0002-4222-1819}{ORCID}) [contributor]
  \item Ellis Hughes \email{ellishughes@live.com} (\href{https://orcid.org/0000-0003-0637-4436}{ORCID}) [contributor]
  \item Thomas Fung \email{thomas.fung.dr@gmail.com} (\href{https://orcid.org/0000-0003-2601-0728}{ORCID}) [contributor]
  \item Reza Norouzian \email{rnorouzian@gmail.com} (\href{https://orcid.org/0000-0002-0531-6886}{ORCID}) [contributor]
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) [contributor]
  \item Quinn White \email{quinnarlise@gmail.com} (\href{https://orcid.org/0000-0001-5399-0237}{ORCID}) [contributor]
  \item David Kane \email{dave.kane@gmail.com} [contributor]
  \item Teun van den Brand \email{t.vandenbrand@nki.nl} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID}) [contributor]
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be} (\href{https://orcid.org/0000-0001-9514-1014}{ORCID}) [contributor]
}

}
\keyword{internal}
