% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equatiomatic-package.R
\docType{package}
\name{equatiomatic-package}
\alias{equatiomatic}
\alias{equatiomatic-package}
\title{equatiomatic: Transform Models into LaTeX Equations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The goal of equatiomatic is to reduce the pain associated with
             writing LaTeX formulas from fitted models. The primary function of
             the package, extract_eq, takes a fitted model object as its input
             and returns the corresponding LaTeX code for the model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datalorax/equatiomatic}
  \item Report bugs at \url{https://github.com/datalorax/equatiomatic/issues}
}

}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

Authors:
\itemize{
  \item Andrew Heiss \email{andrew@andrewheiss.com} (\href{https://orcid.org/0000-0002-3948-3914}{ORCID})
}

Other contributors:
\itemize{
  \item Joshua Rosenberg \email{jmrosenberg@utk.edu} (\href{http://orcid.org/0000-0003-2170-0447}{ORCID}) [contributor]
  \item Jonathan Sidi \email{yonis@metrumrg.com} (\href{https://orcid.org/0000-0002-4222-1819}{ORCID}) [contributor]
}

}
\keyword{internal}
