% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lhs.R
\name{extract_lhs}
\alias{extract_lhs}
\title{Generic function for extracting the left hand side from a model}
\usage{
extract_lhs(model, ...)
}
\arguments{
\item{model}{A fitted model}

\item{\dots}{additional arguments passed to the specific extractor}
}
\description{
Generic function for extracting the left hand side from a model
}
\keyword{internal}
