% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{fix_coef_signs}
\alias{fix_coef_signs}
\title{Deduplicate operators}
\usage{
fix_coef_signs(eq, fix_signs)
}
\arguments{
\item{eq}{String containing a LaTeX equation}

\item{fix_signs}{Logical, defaults to \code{FALSE}. If disabled,
coefficient estimates that are negative are preceded with a "+" (e.g. 5(x)
+ -3(z)). If enabled, the "+ -" is replaced with a "-" (e.g. 5(x) - 3(z)).}
}
\description{
Convert "+ -" to "-"
}
\keyword{internal}
