% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{fix_coef_signs}
\alias{fix_coef_signs}
\title{Normalize operators}
\usage{
fix_coef_signs(eq, fix_signs)
}
\arguments{
\item{eq}{String containing a LaTeX equation}

\item{fix_signs}{Passed from \code{build_tex}}
}
\description{
Convert "+ -" to "-"
}
\keyword{internal}
