% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eq.R
\name{fix_coef_signs}
\alias{fix_coef_signs}
\title{Deduplicate operators}
\usage{
fix_coef_signs(eq)
}
\arguments{
\item{eq}{String containing a LaTeX equation}
}
\description{
Convert "+ -" to "-"
}
\keyword{internal}
