% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.equation}
\alias{format.equation}
\title{Format 'LaTeX' equations}
\usage{
\method{format}{equation}(x, ..., latex = knitr::is_latex_output())
}
\arguments{
\item{x}{'LaTeX' equation built with \code{\link{extract_eq}}}

\item{...}{not used}

\item{latex}{Logical, whether the output is LaTeX or not. The default
value uses {knitr} to determine the current output format.}
}
\value{
A character string with the equation formatted either as proper
LaTeX code, or as a display equation tag (surrounded by \verb{$$...$$}) for R
Markdown or Quarto documents.
}
\description{
Format 'LaTeX' equations built with \code{\link{extract_eq}}.
}
