% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.equation}
\alias{knit_print.equation}
\title{Print 'LaTeX' equations in Rmarkdown environments}
\usage{
\method{knit_print}{equation}(
  x,
  ...,
  tex_packages = "\\\\renewcommand*\\\\familydefault{\\\\rmdefault}"
)
}
\arguments{
\item{x}{'LaTeX' equation built with \code{\link{extract_eq}}}

\item{...}{not used}
}
\description{
Print 'LaTeX' equations built with \code{\link{extract_eq}} nicely in Rmarkdown environments.
}
\keyword{internal}
