% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapply_chr}
\alias{mapply_chr}
\title{Strict versions of \code{base::\link[base]{mapply}}}
\usage{
mapply_chr(...)
}
\arguments{
\item{\dots}{arguments passed to \code{base::\link[base]{mapply}}}
}
\value{
A vector of the corresponding type, \code{chr} = character,
\code{dbl} = double, \code{lgl} = logical, and \code{int} = logical
}
\description{
Return a vector of the corresponding type
}
\keyword{internal}
