% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{preview}
\alias{preview}
\alias{preview.tex}
\title{Preview equation}
\usage{
preview(x, ...)

\method{preview}{tex}(x, ...)
}
\arguments{
\item{x}{LaTeX equation built with \code{build_tex}}

\item{\dots}{arguments passed to \link[texPreview:tex_preview]{texPreview::tex_preview}}
}
\description{
Use \link[texPreview:tex_preview]{texPreview::tex_preview} to preview the final
equation.
}
\details{
To use this method \code{texPreview} must be installed.
}
