% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{preview}
\alias{preview}
\title{Preview equation}
\usage{
preview(eq, ...)
}
\arguments{
\item{eq}{LaTeX equation built with \code{build_tex}}
}
\description{
Use \code{texPreview::\link[texPreview]{tex_preview}} to preview the final
equation.
}
\keyword{internal}
