% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.equation}
\alias{print.equation}
\title{Print LaTeX equations}
\usage{
\method{print}{equation}(x, ...)
}
\arguments{
\item{x}{LaTeX equation built with \code{\link{extract_eq}}}

\item{...}{not used}
}
\description{
Print LaTeX equations built with \code{\link{extract_eq}}.
}
